requirements.txt
langchain
streamlit
python-dotenv
PyPDF2
google-generativeai
langchain_google_genai
-e .